requirements.txt
langchain
langchain-community
langchain-google-genai
pymupdf
python-dotenv
chromadb
langchain-chroma
langchain-openai
fastapi
uvicorn
streamlit
httpx
